+++ crates/ruma-client-api/src/discovery/get_authentication_issuer.rs
@@ -1,6 +1,11 @@
 //! `GET /_matrix/client/*/auth_issuer`
 //!
 //! Get the OpenID Connect Provider that is trusted by the homeserver.
+//!
+//! This endpoint has been replaced by [`get_authorization_server_metadata`] in [MSC2965].
+//!
+//! [`get_authorization_server_metadata`]: super::get_authorization_server_metadata
+//! [MSC2965]: https://github.com/matrix-org/matrix-spec-proposals/pull/2965
 
 pub mod msc2965 {
     //! `MSC2965` ([MSC])
@@ -24,6 +29,7 @@ pub mod msc2965 {
     /// Request type for the `auth_issuer` endpoint.
     #[request(error = crate::Error)]
     #[derive(Default)]
+    #[deprecated = "Replaced by the get_authorization_server_metadata endpoint."]
     pub struct Request {}
 
     /// Request type for the `auth_issuer` endpoint.
@@ -33,6 +39,7 @@ pub mod msc2965 {
         pub issuer: String,
     }
 
+    #[allow(deprecated)]
     impl Request {
         /// Creates a new empty `Request`.
         pub fn new() -> Self {