+++ crates/ruma-client-api/CHANGELOG.md
@@ -1,5 +1,10 @@
 # [unreleased]
 
+Improvements:
+
+- Add support for the authorization server metadata endpoint, according to the
+  latest draft of MSC2965.
+
 # 0.20.1
 
 Bug fixes: